Include Update Feature (#13)
* Refactor get data object for reuse

* Add basic update integration test

* Check for new aliases

* No-op updates not yet working

* Testing apparatus changes to support modification

* Test around update race condition

* Satisfy update integration test

* Add license closes #14

* Add basic auth test

* Satisfy auth test
The header checked is  which can be set via environment variables

* Add auth checks to update test

* Satisfy auth tests for update

* Add configuration of auth to README

* Add meat and potatoes test

* Fix incorrect test
This test assumed an improper request format. This came up when using the notebook, that uses swagger validation on requests and responses. This software could be improved by adding OpenAPI validation on requests on responses.

* Fix update request handling
The request format was incorrect, a data_object key is expected in the request body. In addition, the next_page_token is not included instead of set to null.

* Provide example usage of update mechanism

* Add test to cover alias parsing

* Satisfy multiple colon edge case