Add ElasticSearch domain template (closes #36)
Added a script to spin up an ElasticSearch domain populated with
test data. This is needed to provide an environment with 'fresh'
data. It is also needed so we can implement data bundle
functionality experimentally without affecting dss-azul-commons.

Instead of spinning up a new instance for each test, it should be
enough to have a single ES instance always running that Travis
can then wipe and populate as necessary for testing.