Fixes `obj_to_azul` field leakage (fixes #111)
Also closes #110.

This commit fixes attribute leakage in `obj_to_azul` when making
an UpdateDataObject request by (1) modifying the UpdateDataObject
endpoint to making a partial document update request to Azul
instead of replacing the document entirely and (2) correcting the
behavior of `obj_to_azul` to remove fields that shouldn't be there.