Replace boto with boto3 (closes #73)
This commit updates app.py and specified dependencies to use
boto3 over boto, which allows us to improve our Python 3.x support.
In the process of doing so, "manual" calls to ElasticSearch have
been replaced with the ElasticSearch Python bindings, which provide
more robust logging and error handling.

The provisioning script is also updated to use boto3 and the
ElasticSearch bindings.